requirements.txt
defusedxml==0.5.0
Django==2.1.5
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-extra-views==0.12.0
django-filter==2.1.0
django-floppyforms==1.7.0
django-import-export==1.2.0
django-recaptcha==2.0.0
django-reversion==3.0.2
django-sendfile==0.3.11
django-simple-captcha==0.5.10
pytz==2016.4
lxml==3.8.0
python-memcached==1.59
requests==2.21.0
zeep==2.1.1
django-countries==5.3.2
